LayoutTests/fast/dom/form-elements-collection-idempotence-expected.txt

This test verifies that HTMLFormElement.elements returns the same collection when called repeatedly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS formElement.elements is formElement.elements
PASS formElement.elements === formElement.elements is true
PASS successfullyParsed is true

TEST COMPLETE
